refactor: Removed puppeteer and used http request with html parser instead
