+++ lib/src/ingredient_output_generator.dart
@@ -1,0 +1,72 @@
+import 'package:flutter/cupertino.dart' show visibleForTesting;
+import 'package:intl/intl.dart' show NumberFormat;
+
+import 'recipe_models.dart' show RecipeCollectionResult, Recipe, Ingredient;
+
+/// Creates different representations of the passed [Recipe]s.
+///
+/// The [RecipeCollectionResult] contains strings with different representations
+///
+RecipeCollectionResult createCollectionResultFromRecipes(
+  List<Recipe> recipes,
+) {
+  var mergedIngredients = mergeIngredients(
+    recipes.expand<Ingredient>((recipe) => recipe.ingredients).toList(),
+  );
+  var ingredientsSortedByAmount = mergedIngredients
+    ..sort((a, b) => b.amount.compareTo(a.amount));
+
+  return RecipeCollectionResult(
+    resultSortedByAmount:
+        _convertIngredientsToString(ingredientsSortedByAmount),
+  );
+}
+
+@visibleForTesting
+
+/// Merges the [Ingredient]s in the passed list to a list of unique
+/// [Ingredient]s.
+///
+/// [Ingredient]s with the same name and unit are merged so that the amount is
+/// added together e.g. 4 g Sugar and 8 g Sugar results in 12 g Sugar.
+List<Ingredient> mergeIngredients(List<Ingredient> ingredients) {
+  var mergedIngredients = <Ingredient>[];
+
+  for (var ingredient in ingredients) {
+    var index = mergedIngredients.indexWhere(
+      (element) =>
+          ingredient.name == element.name && ingredient.unit == element.unit,
+    );
+
+    if (index == -1) {
+      mergedIngredients.add(ingredient);
+    } else {
+      var amount = mergedIngredients[index].amount + ingredient.amount;
+      mergedIngredients[index] = Ingredient(
+        amount: amount,
+        unit: ingredient.unit,
+        name: ingredient.name,
+      );
+    }
+  }
+
+  return mergedIngredients;
+}
+
+String _convertIngredientsToString(List<Ingredient> ingredients) {
+  var result = "";
+  for (var ingredient in ingredients) {
+    if (ingredient.amount > 0) {
+      var formatter = NumberFormat()
+        ..minimumFractionDigits = 0
+        ..maximumFractionDigits = 2;
+
+      result += "${formatter.format(ingredient.amount)} ";
+    }
+    if (ingredient.unit.isNotEmpty) {
+      result += "${ingredient.unit} ";
+    }
+    result += "${ingredient.name}\n";
+  }
+  return result;
+}
