+++ lib/src/recipe_controller.dart
@@ -1,118 +1,40 @@
-import 'dart:io';
+import 'package:html/dom.dart' show Document;
+import 'package:html/parser.dart' show parse;
+import 'package:http/http.dart' as http;
 
-import 'package:flutter/cupertino.dart';
-import 'package:intl/intl.dart';
-import 'package:puppeteer/puppeteer.dart';
-
+import 'recipe-scripts/kptncook.dart';
 import 'recipe_models.dart';
 
-const _scriptBasePath = 'lib/src/recipe-scripts/';
-
-final Map<String, String> _scriptPathMap = {
-  'mobile.kptncook.com': 'kptncook.js',
+final Map<String, Recipe Function(Document, int)> _recipeParseMethodMap = {
+  'mobile.kptncook.com': parseKptnCookRecipe,
 };
 
-/// Collects ingredients from passed [recipes].
+/// Collects recipes from the websites in the passed [recipeInfos].
 ///
-/// Optional [language] is set as browser language.
-Future<RecipeCollectionResult> collectIngredients(
-  List<RecipeData> recipes, [
-  String? language,
-]) async {
-  var headers = <String, String>{};
-
-  if (language != null) {
-    headers['Accept-Language'] = 'language';
-  }
-
-  // Launch browser and open new tab
-  var browser = await puppeteer.launch();
-  var page = await browser.newPage();
-  await page.setExtraHTTPHeaders(headers);
-
-  // Call scripts for each recipe
+/// For each [RecipeInfo] a http request is made to the website containing the
+/// recipe. From there the recipe is parsed and adjusted to the amount of
+/// servings. The list of the parsed [Recipe]s is returned.
+Future<List<Recipe>> collectRecipes(List<RecipeInfo> recipeInfos) async {
   var results = <Recipe>[];
 
-  for (var recipe in recipes) {
-    // Get recipe script
-    // TODO: Give not-supported feedback when host is not in map
-    var scriptPath = _scriptBasePath + (_scriptPathMap[recipe.url.host] ?? '');
-    var scriptFilePath =
-        Uri(path: scriptPath).toFilePath(windows: Platform.isWindows);
-    var scriptFile = await File(scriptFilePath).readAsString();
+  var client = http.Client();
 
-    // Go to URL and execute script
-    await page.goto(recipe.url.toString(), wait: Until.networkIdle);
-    var result = await page.evaluate(scriptFile, args: [recipe.servings]);
-
-    // Parse result from json and add to list
-    results.add(Recipe.fromJson(result));
+  for (var recipe in recipeInfos) {
+    var resultRecipe = await _collectRecipe(client, recipe);
+    results.add(resultRecipe);
   }
 
-  await browser.close();
-  return _createCollectionResultFromRecipes(results);
+  client.close();
+
+  return results;
 }
 
-RecipeCollectionResult _createCollectionResultFromRecipes(
-  List<Recipe> recipes,
-) {
-  var mergedIngredients = mergeIngredients(
-    recipes.expand<Ingredient>((recipe) => recipe.ingredients).toList(),
-  );
-  var ingredientsSortedByAmount = mergedIngredients
-    ..sort((a, b) => b.amount.compareTo(a.amount));
+Future<Recipe> _collectRecipe(http.Client client, RecipeInfo recipe) async {
+  var response = await client.get(recipe.url);
 
-  return RecipeCollectionResult(
-    resultSortedByAmount:
-        _convertIngredientsToString(ingredientsSortedByAmount),
-  );
-}
+  var document = parse(response.body);
 
-@visibleForTesting
+  var parseMethod = _recipeParseMethodMap[recipe.url.host];
 
-/// Merges the [Ingredient]s in the passed list to a list of unique
-/// [Ingredient]s.
-///
-/// [Ingredient]s with the same name and unit are merged so that the amount is
-/// added together e.g. 4 g Sugar and 8 g Sugar results in 12 g Sugar.
-List<Ingredient> mergeIngredients(List<Ingredient> ingredients) {
-  var mergedIngredients = <Ingredient>[];
-
-  for (var ingredient in ingredients) {
-    var index = mergedIngredients.indexWhere(
-      (element) =>
-          ingredient.name == element.name && ingredient.unit == element.unit,
-    );
-
-    if (index == -1) {
-      mergedIngredients.add(ingredient);
-    } else {
-      var amount = mergedIngredients[index].amount + ingredient.amount;
-      mergedIngredients[index] = Ingredient(
-        amount: amount,
-        unit: ingredient.unit,
-        name: ingredient.name,
-      );
-    }
-  }
-
-  return mergedIngredients;
-}
-
-String _convertIngredientsToString(List<Ingredient> ingredients) {
-  var result = "";
-  for (var ingredient in ingredients) {
-    if (ingredient.amount > 0) {
-      var formatter = NumberFormat()
-        ..minimumFractionDigits = 0
-        ..maximumFractionDigits = 2;
-
-      result += "${formatter.format(ingredient.amount)} ";
-    }
-    if (ingredient.unit.isNotEmpty) {
-      result += "${ingredient.unit} ";
-    }
-    result += "${ingredient.name}\n";
-  }
-  return result;
+  return parseMethod!.call(document, recipe.servings);
 }
