+++ pubspec.yaml
@@ -9,9 +9,10 @@ dependencies:
   flutter:
     sdk: flutter
   freezed_annotation: ^2.2.0
+  html: ^0.15.1
+  http: ^0.13.5
   intl: ^0.18.0
   json_annotation: ^4.7.0
-  puppeteer: ^2.17.0
 
 dev_dependencies:
   flutter_test:
